working on getting backend to accept new user

+++ app/controllers/api/users_controller.rb
@@ -12,20 +12,16 @@ class Api::UsersController < ApplicationController
         # 12 might need additional logic to populate f_name, l_name, and username
         if @user.save
             login!(@user)
-            render :show # 15 # this routes should be "api/users/show"
+            render :show
         else
-            render json: @user.errors.full_messages, status: 422
+            render json: @user.errors.full_messages, status: 401
         end
 
     end
 
-    def index
-        @user = User.find(params[:id])
-    end
 
     def show
         @user = User.find(params[:id])
-        # 28 render :show # this is where well link to the jbuilder that gets a user's profile info
     end
 
     # def edit
@@ -34,17 +30,20 @@ class Api::UsersController < ApplicationController
     def update
         @user = User.find(params[:id])
         if @user && @user.update_attributes(user_params)
-            render "api/users/show"
+            render :show
         else
-            render jason: @user.errors.full_messages # not sure what status to assign
+            render jason: @user.errors.full_messages, status: 401
         end
     end
 
 
     def destroy
-        user = User.find(params[:id])
-        user.destroy
-        render "api/session/new" # not sure if this is correct
+        @user = User.find(params[:id])
+        if @user
+            user.destroy
+        else
+            render json: ['User not found']
+        end
     end
 
     private
